chore(rust): split decoder::root into limits and root01
